requirements.txt
torchvision
torch
Pillow
opencv-python
matplotlib
segmentation-models-pytorch
torchmetrics
fastapi
uvicorn
pydantic
albumentations
python-multipart